requirements.txt
Package                Version
---------------------- ---------
absl-py                0.10.0
astor                  0.8.1
astunparse             1.6.3
attrs                  20.2.0
Automat                20.2.0
cachetools             4.1.1
certifi                2020.6.20
cffi                   1.14.3
chardet                3.0.4
constantly             15.1.0
cryptography           3.1.1
cssselect              1.1.0
gast                   0.2.2
google-auth            1.21.3
google-auth-oauthlib   0.4.1
google-pasta           0.2.0
grpcio                 1.32.0
h5py                   2.10.0
hyperlink              20.0.1
idna                   2.10
importlib-metadata     2.0.0
incremental            17.5.0
itemadapter            0.1.0
itemloaders            1.0.3
jmespath               0.10.0
joblib                 0.16.0
Keras                  2.4.3
Keras-Applications     1.0.8
Keras-Preprocessing    1.1.2
lxml                   4.5.2
Markdown               3.2.2
mock                   4.0.2
mouse                  0.7.1
mysqlclient            2.0.1
numpy                  1.16.1
oauthlib               3.1.0
opt-einsum             3.3.0
parsel                 1.6.0
Pillow                 7.2.0
pip                    20.2.3
Protego                0.1.16
protobuf               3.13.0
pyasn1                 0.4.8
pyasn1-modules         0.2.8
pycparser              2.20
PyDispatcher           2.0.5
PyHamcrest             2.0.2
PyMySQL                0.10.1
pyOpenSSL              19.1.0
PyYAML                 5.3.1
queuelib               1.5.0
redis                  3.5.3
requests               2.24.0
requests-oauthlib      1.3.0
rsa                    4.6
scikit-learn           0.23.2
scipy                  1.4.1
Scrapy                 2.3.0
scrapy-redis           0.6.8
selenium               3.141.0
service-identity       18.1.0
setuptools             49.6.0
six                    1.15.0
SQLAlchemy             1.3.20
tensorboard            1.15.0
tensorboard-plugin-wit 1.7.0
tensorflow             1.15.2
tensorflow-estimator   1.15.1
termcolor              1.1.0
Theano                 1.0.5
threadpoolctl          2.1.0
Twisted                20.3.0
urllib3                1.25.10
w3lib                  1.22.0
Werkzeug               1.0.1
wheel                  0.35.1
wrapt                  1.12.1
zipp                   3.2.0
zope.interface         5.1.0
